fix: error on undefined
DAG-CBOR only supports the CBOR value `null` and not `undefined`, hence
error on `undefined`. This matches the JS and Go implementation.

This could be seen as a breaking change, but the point is, that it now
follows the spec (and other implementations) more closely, so it can
be considered a bug fix to prevent incompatible data.

BREAKING CHANGE: undefined is disallowed now (matching the DAG-CBOR spec)