fix: validate integers before encoding. (#15)
Integers in CBOR range from -2^64 to 2^64-1 inclusive. This means that we can have negative
integers that are bigger than the usual minimum. Internally we use i128 for the integers,
but we need to make sure that if integers encoded that they fall into the range CBOR
specifies. This commit introduces such a check.

Fixes #14.