fix: add information about limitation and examples
Serde derives for internally tagged and untagged unions are not supported.
The reason is, that this library tries to make it impossible to deserialize
a CID accidentally into bytes.

This limiation is now documented and examples on how to work around it are
also provided. An additional test that fails when trying to deserialize a
CID into bytes was also added.