feat: introduce no-cid-as-bytes feature
This commit changes the default behaviour of the deserializer. It's now
possible to deserialize CIDs into bytes. This is needed in order to have
the Serde derives for internally tagged and untagged enums working correctly.

If you need the current behaviour of erroring at run-time in case a CID
is tried to be deserialized as bytes, you can enable the `no-cid-as-bytes`
feature.

BREAKING CHANGE: CIDs may be deserialized into bytes. In order to restore
the original behaviour, enable the `no-cid-as-bytes` feature.