feat: allow serializer to be used with serde-transcode
Prior to this change, the `Serializer` was private. With this change it's now
possible to use this library with `serde-transcode` [1].

The handling of the map key order needed to be changed. We were using `collect_map`
in order to make sure maps have the correct order. But that is only called when a
`BTreeMap` or `HashMap` is used [2]. `serde-transcode` calls `serialize_map`
directly. Therefore the logic to do the sorting was moved into the `SerializeMap`
implementation. This way it's used by all known cases.

[1]: https://crates.io/crates/serde-transcode
[2]: https://github.com/serde-rs/serde/blob/00c4b0cef80557c33fbcd75fcc70dc034720b4df/serde/src/ser/impls.rs#L430-L487

Closes #26.