fix: make sure sequences are always bound
With the `Serializer` being public and being able to being used with
serde-transcode [1], we need to make sure that always valid DAG-CBOR
is generated, this means that sequences are always bound ones, where
the size is known beforehand. If the input is an unbound sequence,
then the data is buffered until the total size was determined.

[1]: https://crates.io/crates/serde-transcode