feat: implement Codec trait (#28)
The `Codec` trait from `ipld-core` allows unified access for encoding,
decoding and extracting links of encoded IPLD data independent of the
codec. It only works when the `no-cid-as-bytes` feature is disabled,
because the link extraction code is based on Serde.

This commit also makes the `Deserializer` public, so that it can be
used for `ipld-extract-links`. Having it public should also help with
issue #25.