feat: add de::from_reader_once
The existing de::from_reader will fail if there is trailing data at
the end of the reader. This leads to patterns such as passing a
cursor in as the reader, allowing it to fail with
DecodeError::TrailingData, checking how much data was used before
the error, creating a slice containing only the portion before the
trailing data, and reparsing[1].

This new function allows the user to decode a concatenated stream
of values without having to double-parse to determine the length.

1. https://docs.rs/crate/bluesky-firehose-stream/0.1.1/source/src/frame.rs#84